Add Netlify headers configuration
This is only used if the site's config has the necessary settings, so
can ignored when not using Netlify.

Note that the exampleSite in this repository is configured for Netlify.
Also note that the header currently configured is a
Content-Security-Policy compatible with current ESR versions of Firefox.
Eventually (as ESR browsers are replaced with newer versions) it will
be possible to use a 'tighter' CSP by default.